requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pygments==2.17.2
PyJWT==2.8.0
rest-framework-roles==1.0.6
sqlparse==0.5.0
